Make mention's uniquely hashable
A u32 has been added to Mention to indicated whether the Mentionable to
which it refers is the nth occurance. Two identical Mention's we're
previously indistinguishable clones. With the addition of the
`occurance` field each Mention is uniquely identifiable.

Using the Mention hash for the HTML element's id, one can link to a
specific Mention.