Remove asset attribution bubble
Was added when info bubbles where few.
Now there can be three bubbles with lots of info to digest.
The attribution info is now partially duplicated in the support bubble.
The exact author of a transcript isn't that useful to the reader.
The attribution data is still in the markdown source.
There is now, unfortunately, no visual reward for contributing a transcript.