Add `mirrors` & `prev-filenames` for 2016-11-18
Introduce `source.mirrors` as a list of locations to provide to readers,
and to programmatically compare assets parity between projects.

Introduce `prev-filenames` as a record of past names of an asset. When
changing an assets filename keeping a record of previous names allows
for provision of URL redirection, or disambiguation suggestions on pages
which takeup that old name, such as mention pages.