Add `check` subcmd + refactor ./cmd to ./internal
The `check` sub command is a work in progress that crawls other Ray Peat
archives and recommends changes to Ray Peat Rodeo.

The refactor is largely to facilitate the sharing of code with the
`check` package, code that used to be in the `main` package. It's also a
little bit tidier overall.