Change popup files to use slugs for dir/filenames
After upgrading the GitHub workflow to use upload-artifact@v4, it now
refuses to accept colons within filenames (in order to support NTFS).
Mentions can be URLs (i.e. "https://...") which include colons.
Forcing mentions to be a slug compliant introduces a source of conflict
as many different names can be slugified into the same result. But
we have a conflict detection that panics if that happens, and it's not
happening at the moment at all. So fine.