Rewrite SSG in Node and frontend in Lit & Vite
After deciding on Web Components, Go types had to be duplicated in JSDoc
syntax for the JS client code. Although rewriting the static site
generation in Node makes SSG slower for a total build, the developer
experience can actually be faster because reloading Web Components is
fast.

Initially native Web Components were working well, but the value of
reactive state mounted. Lit Elements add reactive state to Web Component
standards, so that, unlike React, it's a small dependency, and requires
no build step.

The client UI itself is now a two sidebar layout, both attached to the
left of the viewport, each scrollable. One for filters, and one for
results, with content being loading in the page body.

Pagefind performance improves a lot with this commit following the
adoption of an IntersectionObserver for loading data as discussed in
cloudcannon/pagefind#371. Rurther reductions to initalisation times
to follow with the proposed fixes in that issue.

Client JS now uses modules (as opposed to synchronous plain JS) in the
browser. This approach initially broke support for some older browsers,
specifically iOS 16.3 and it's lack of support for import maps. To solve
this Vite is now being used to compile client code in a way that's
compatible with older browsers.